postHook on changes, part I
View3D now has the information for only calling postHooks when the frame has changed. Right now it is only based on camera movement and works, but TODO: when the rendering changes because of time or hash changes, this CANNOT be detected by design. TODO: propagate frameIsNew on CallRender3D or something like that.