Fixed the CUDA compiler
Using the add_compile_options function to add /utf-8 to the compile options for MSVC  breaks the nvcc compiler since the /utf-8 flag is added to the Command Line, resulting in the following error: nvcc fatal   : A single input file is required for a non-link phase when an outputfile is specified
Therefore the add_compile_options call have been removed and the /utf-8 was added to the COMMON_CXX_FLAGS.

Upgraded the minimum CUDA version to 11 sind it is the first version to support c++17, for both the protein_cuda and the molecularmaps plugin.

Upgraded the molecularmaps plugin to use the new camera.